+++ packages/native/src/components/containers/Input/index.tsx
@@ -7,7 +7,7 @@ import { inputConfig } from './config'
 export const Input = makeComponent(TextInput, inputConfig)
 
 export const pullInputProps = makePropPuller(inputConfig)
-export const getInputStyle = makeStyleGetter(inputConfig)
+export const getInputStyles = makeStyleGetter(inputConfig)
 
 export * from './types'
 export * from './config'