+++ packages/web/src/components/content/Text/index.tsx
@@ -5,7 +5,7 @@ import { textConfig } from './config'
 export const Text = makeComponent('span', textConfig)
 
 export const pullTextProps = makePropPuller(textConfig)
-export const getTextStyle = makeStyleGetter(textConfig)
+export const getTextStyles = makeStyleGetter(textConfig)
 
 export * from './types'
 export * from './config'