+++ packages/web/src/components/controls/Input/core/Display.tsx
@@ -1,9 +1,9 @@
 import { forwardRef, MutableRefObject } from 'react'
 
-import { DisplayInputProps } from './types'
 import { makeComponent } from '../../../../utils'
 import { rowConfig } from '../../../containers'
-import { inputConfig } from '../styles/config'
+import { DisplayInputProps } from './types'
+import inputConfig from './config'
 
 const Container = makeComponent(
   'div',