+++ packages/web/src/stories/FormInput.stories.tsx
@@ -3,7 +3,7 @@ import { ComponentStory, ComponentMeta } from '@storybook/react'
 import { useState } from 'react'
 import { ThemeProvider } from 'styled-components'
 
-import { FormInput } from '../components'
+import { Column, FormInput } from '../components'
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
@@ -21,25 +21,28 @@ export default {
     //   options: ['Primary', 'Success', 'Danger', 'Warning'],
     // },
     // radius: { control: 'select', options: ['sm', 'md', 'lg', 'pill'] },
-    // value: { table: { disable: true } },
+    errorText: { control: 'text' },
     // setValue: { table: { disable: true } },
   },
 } as ComponentMeta<typeof FormInput>
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof FormInput> = () => {
+const Template: ComponentStory<typeof FormInput> = (args) => {
   const [value, setValue] = useState('')
   return (
     <ThemeProvider theme={makeTheme({})}>
-      <FormInput
-        fullWidth
-        maxUnitWidth={35}
-        value={value}
-        setValue={setValue}
-        label="My test input."
-        placeholder="Type something here..."
-        helperText="This is a helper text for the input."
-      />
+      <Column unitWidth={40}>
+        <FormInput
+          fullWidth
+          maxUnitWidth={35}
+          label="My test input."
+          placeholder="Type something here..."
+          helperText="This is a helper text for the input."
+          {...args}
+          value={value}
+          onChange={(e) => setValue(e.currentTarget.value)}
+        />
+      </Column>
     </ThemeProvider>
   )
 }